+++ app/src/main/java/com/example/nearfall/SignUpFragment.java
@@ -5,10 +5,8 @@ import android.app.DatePickerDialog;
 import android.content.Context;
 import android.database.SQLException;
 import android.os.Bundle;
-
 import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
-
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -23,11 +21,9 @@ import android.widget.Button;
 import android.widget.RadioButton;
 import android.widget.TextView;
 import android.widget.Toast;
-
 import com.example.nearfall.User.HashedPassword;
 import com.example.nearfall.User.User;
 import com.example.nearfall.User.UserManager;
-
 import java.security.NoSuchAlgorithmException;
 import java.util.Calendar;
