+++ app/src/main/java/com/example/nearfall/Tutorial3Fragment.java
@@ -1,10 +1,8 @@
 package com.example.nearfall;
 
 import android.os.Bundle;
-
 import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
