+++ app/src/main/java/com/example/nearfall/Tutorial8Fragment.java
@@ -1,7 +1,5 @@
 package com.example.nearfall;
 
-import android.content.Context;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -13,9 +11,6 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 public class Tutorial8Fragment extends Fragment implements View.OnClickListener{
-    public static final String MyPREFERENCES = "MyPrefs" ;
-    public static final String Purpose = "purposeKey";
-    SharedPreferences sharedpreferences;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
